+++ src/models/conversation.js
@@ -140,6 +140,7 @@ class Conversation extends Container {
     } : options;
     messageConfig.clientId = this.clientId;
     messageConfig.conversationId = this.id;
+    messageConfig._loadType = 'websocket'; // treat this the same as a websocket loaded object
 
     return new ConversationMessage(messageConfig);
   }